+++ system/modules/efg/dca/tl_module.php
@@ -319,7 +319,7 @@ class tl_module_efg extends \Backend
 
 			// all forms marked to store data
 			$objFields = \Database::getInstance()->prepare("SELECT f.id,f.title,f.alias,f.formID,ff.type,ff.name,ff.label FROM tl_form f, tl_form_field ff WHERE (f.id=ff.pid) AND storeFormdata=? ORDER BY title")
-				->executeUncached('1');
+				->execute('1');
 
 			while ($objFields->next())
 			{